refactor(app): extract dashboard heatmap tile
