+++ packages/app/src/DBDashboardPage/DBDashboardPage.tsx
@@ -33,7 +33,6 @@ import {
 } from '@hyperdx/common-utils/dist/guards';
 import {
   AlertState,
-  BuilderChartConfigWithDateRange,
   ChartConfigWithDateRange,
   DashboardContainer as DashboardContainerSchema,
   DashboardFilter,
@@ -46,7 +45,6 @@ import {
   SearchConditionLanguage,
   SourceKind,
   SQLInterval,
-  TSource,
 } from '@hyperdx/common-utils/dist/types';
 import {
   ActionIcon,
@@ -60,8 +58,6 @@ import {
   Menu,
   Modal,
   Paper,
-  Popover,
-  Portal,
   Stack,
   Text,
   Tooltip,
@@ -82,7 +78,6 @@ import {
   IconPlayerPlay,
   IconPlus,
   IconRefresh,
-  IconSearch,
   IconSquaresDiagonal,
   IconTags,
   IconTrash,
@@ -120,9 +115,6 @@ import useDashboardContainers, {
 import { calculateNextTilePosition, makeId } from '@/utils/tilePositioning';
 
 import ChartContainer from '@/components/charts/ChartContainer';
-import DBHeatmapChart, {
-  toHeatmapChartConfig,
-} from '@/components/DBHeatmapChart';
 import { DBPieChart } from '@/components/DBPieChart';
 import DBSqlRowTableWithSideBar from '@/components/DBSqlRowTableWithSidebar';
 import OnboardingModal from '@/components/OnboardingModal';
@@ -135,11 +127,7 @@ import { useDashboardRefresh } from '@/hooks/useDashboardRefresh';
 import useTileSelection from '@/hooks/useTileSelection';
 import { useBrandDisplayName } from '@/theme/ThemeProvider';
 import { parseAsJsonEncoded, parseAsStringEncoded } from '@/utils/queryParsers';
-import {
-  buildEventsSearchUrl,
-  buildTableRowSearchUrl,
-  DEFAULT_CHART_CONFIG,
-} from '@/ChartUtils';
+import { buildTableRowSearchUrl, DEFAULT_CHART_CONFIG } from '@/ChartUtils';
 import { useConnections } from '@/connection';
 import { useDashboard } from '@/dashboard';
 import DashboardFilters from '@/DashboardFilters';
@@ -159,137 +147,10 @@ import { FormatTime } from '@/useFormatTime';
 import { getMetricTableName } from '@/utils';
 import { useZIndex, ZIndexContext } from '@/zIndex';
 
+import { HeatmapTile } from './HeatmapTile';
+
 import 'react-grid-layout/css/styles.css';
 import 'react-resizable/css/styles.css';
-
-function HeatmapTile({
-  keyPrefix,
-  chartId,
-  title,
-  toolbar,
-  queriedConfig,
-  source,
-  dateRange,
-}: {
-  keyPrefix: string;
-  chartId: string;
-  title: React.ReactNode;
-  toolbar: React.ReactNode[];
-  queriedConfig: BuilderChartConfigWithDateRange;
-  source: TSource | undefined;
-  dateRange: [Date, Date];
-}) {
-  const { heatmapConfig, scaleType } = toHeatmapChartConfig(queriedConfig);
-
-  const [clickPos, setClickPos] = useState<{ x: number; y: number } | null>(
-    null,
-  );
-  const containerRef = useRef<HTMLDivElement>(null);
-
-  const eventDeltasUrl = useMemo(() => {
-    if (!source) return null;
-    const url = buildEventsSearchUrl({
-      source,
-      config: queriedConfig,
-      dateRange,
-    });
-    if (!url) return null;
-    const separator = url.includes('?') ? '&' : '?';
-    return `${url}${separator}mode=delta`;
-  }, [source, queriedConfig, dateRange]);
-
-  const handleClick = useCallback(
-    (e: React.MouseEvent<HTMLDivElement>) => {
-      if (!eventDeltasUrl) return;
-      const rect = containerRef.current?.getBoundingClientRect();
-      if (!rect) return;
-      setClickPos({ x: e.clientX - rect.left, y: e.clientY - rect.top });
-    },
-    [eventDeltasUrl],
-  );
-
-  const dismiss = useCallback(() => setClickPos(null), []);
-
-  return (
-    <div
-      ref={containerRef}
-      style={{ position: 'relative', width: '100%', height: '100%' }}
-      onClick={handleClick}
-    >
-      <DBHeatmapChart
-        key={`${keyPrefix}-${chartId}`}
-        title={title}
-        toolbarPrefix={toolbar}
-        config={heatmapConfig}
-        scaleType={scaleType}
-        showLegend
-      />
-      {clickPos != null && eventDeltasUrl != null && (
-        <>
-          <Portal>
-            <div
-              style={{
-                position: 'fixed',
-                top: 0,
-                left: 0,
-                right: 0,
-                bottom: 0,
-                zIndex: 199,
-              }}
-              onClick={e => {
-                e.stopPropagation();
-                e.preventDefault();
-                dismiss();
-              }}
-              onMouseDown={e => e.stopPropagation()}
-            />
-          </Portal>
-          <Popover
-            opened
-            onChange={opened => {
-              if (!opened) dismiss();
-            }}
-            position="bottom-start"
-            offset={4}
-            withinPortal
-            closeOnEscape
-            withArrow
-            shadow="md"
-          >
-            <Popover.Target>
-              <div
-                style={{
-                  position: 'absolute',
-                  left: clickPos.x,
-                  top: clickPos.y,
-                  width: 1,
-                  height: 1,
-                  pointerEvents: 'none',
-                }}
-              />
-            </Popover.Target>
-            <Popover.Dropdown
-              p="xs"
-              onClick={e => e.stopPropagation()}
-              onMouseDown={e => e.stopPropagation()}
-            >
-              <Link
-                data-testid="heatmap-view-event-deltas-link"
-                href={eventDeltasUrl}
-                onClick={dismiss}
-              >
-                <Group gap="xs">
-                  <IconSearch size={16} />
-                  View in Event Deltas
-                </Group>
-              </Link>
-            </Popover.Dropdown>
-          </Popover>
-        </>
-      )}
-    </div>
-  );
-}
 
 const ReactGridLayout = WidthProvider(RGL);
